+++ src/main/java/org/example/lv_backend/entity/UserSimId.java
@@ -1,5 +1,6 @@
 package org.example.lv_backend.entity;
 
+import jakarta.persistence.Entity;
 import lombok.*;
 
 import java.io.Serializable;
@@ -9,6 +10,6 @@ import java.io.Serializable;
 @AllArgsConstructor
 @EqualsAndHashCode
 public class UserSimId implements Serializable {
-    private Integer userA;
-    private Integer userB;
+    private Long userA;
+    private Long userB;
 }
